+++ src/devices/boiler.h
@@ -55,11 +55,13 @@ class Boiler : public EMSdevice {
     uint8_t mqtt_format_;            // single, nested or ha
     bool    changed_ = false;
 
-    static constexpr uint8_t EMS_TYPE_UBAParameterWW  = 0x33;
-    static constexpr uint8_t EMS_TYPE_UBAFunctionTest = 0x1D;
-    static constexpr uint8_t EMS_TYPE_UBAFlags        = 0x35;
-    static constexpr uint8_t EMS_TYPE_UBASetPoints    = 0x1A;
-    static constexpr uint8_t EMS_TYPE_UBAParameters   = 0x16;
+    static constexpr uint8_t EMS_TYPE_UBAParameterWW     = 0x33;
+    static constexpr uint8_t EMS_TYPE_UBAFunctionTest    = 0x1D;
+    static constexpr uint8_t EMS_TYPE_UBAFlags           = 0x35;
+    static constexpr uint8_t EMS_TYPE_UBASetPoints       = 0x1A;
+    static constexpr uint8_t EMS_TYPE_UBAParameters      = 0x16;
+    static constexpr uint8_t EMS_TYPE_UBAParametersPlus  = 0xE6;
+    static constexpr uint8_t EMS_TYPE_UBAParameterWWPlus = 0xEA;
 
     static constexpr uint8_t EMS_BOILER_SELFLOWTEMP_HEATING = 20; // was originally 70, changed to 30 for issue #193, then to 20 with issue #344
 
@@ -68,7 +70,7 @@ class Boiler : public EMSdevice {
     uint8_t wWSelTemp_       = EMS_VALUE_UINT_NOTSET; // Warm Water selected temperature
     uint8_t wWCircPump_      = EMS_VALUE_BOOL_NOTSET; // Warm Water circulation pump available
     uint8_t wWCircPumpMode_  = EMS_VALUE_UINT_NOTSET; // Warm Water circulation pump mode
-    uint8_t wWCircPumpType_  = EMS_VALUE_BOOL_NOTSET; // Warm Water circulation pump type
+    uint8_t wWChargeType_    = EMS_VALUE_BOOL_NOTSET; // Warm Water charge type (pump or 3-way-valve)
     uint8_t wWDisinfectTemp_ = EMS_VALUE_UINT_NOTSET; // Warm Water disinfection temperature to prevent infection
     uint8_t wWComfort_       = EMS_VALUE_UINT_NOTSET; // WW comfort mode
 
@@ -117,20 +119,22 @@ class Boiler : public EMSdevice {
     uint8_t  wWRecharging_    = EMS_VALUE_BOOL_NOTSET;   // Warm Water recharge on/off
     uint8_t  wWTemperatureOK_ = EMS_VALUE_BOOL_NOTSET;   // Warm Water temperature ok on/off
     uint8_t  wWCurFlow_       = EMS_VALUE_UINT_NOTSET;   // Warm Water current flow temp in l/min
+    uint8_t  wWType_          = EMS_VALUE_UINT_NOTSET;   // 0-off, 1-flow, 2-flowbuffer, 3-buffer, 4-layered buffer
 
     // UBATotalUptime
     uint32_t UBAuptime_ = EMS_VALUE_ULONG_NOTSET; // Total UBA working hours
 
     // UBAParameters
-    uint8_t heating_temp_ = EMS_VALUE_UINT_NOTSET; // Heating temperature setting on the boiler
-    uint8_t pump_mod_max_ = EMS_VALUE_UINT_NOTSET; // Boiler circuit pump modulation max. power %
-    uint8_t pump_mod_min_ = EMS_VALUE_UINT_NOTSET; // Boiler circuit pump modulation min. power
-    uint8_t burnPowermin_ = EMS_VALUE_UINT_NOTSET;
-    uint8_t burnPowermax_ = EMS_VALUE_UINT_NOTSET;
-    int8_t  boilTemp_off_ = EMS_VALUE_INT_NOTSET;
-    int8_t  boilTemp_on_  = EMS_VALUE_INT_NOTSET;
-    uint8_t burnPeriod_   = EMS_VALUE_UINT_NOTSET;
-    uint8_t pumpDelay_    = EMS_VALUE_UINT_NOTSET;
+    uint8_t heating_activated_ = EMS_VALUE_BOOL_NOTSET; // Heating activated on the boiler
+    uint8_t heating_temp_      = EMS_VALUE_UINT_NOTSET; // Heating temperature setting on the boiler
+    uint8_t pump_mod_max_      = EMS_VALUE_UINT_NOTSET; // Boiler circuit pump modulation max. power %
+    uint8_t pump_mod_min_      = EMS_VALUE_UINT_NOTSET; // Boiler circuit pump modulation min. power
+    uint8_t burnPowermin_      = EMS_VALUE_UINT_NOTSET;
+    uint8_t burnPowermax_      = EMS_VALUE_UINT_NOTSET;
+    int8_t  boilTemp_off_      = EMS_VALUE_INT_NOTSET;
+    int8_t  boilTemp_on_       = EMS_VALUE_INT_NOTSET;
+    uint8_t burnPeriod_        = EMS_VALUE_UINT_NOTSET;
+    uint8_t pumpDelay_         = EMS_VALUE_UINT_NOTSET;
 
     // UBASetPoint
     uint8_t setFlowTemp_  = EMS_VALUE_UINT_NOTSET; // boiler setpoint temp
@@ -151,6 +155,8 @@ class Boiler : public EMSdevice {
     void process_UBAMonitorSlow(std::shared_ptr<const Telegram> telegram);
     void process_UBAMonitorSlowPlus(std::shared_ptr<const Telegram> telegram);
     void process_UBAMonitorSlowPlus2(std::shared_ptr<const Telegram> telegram);
+    void process_UBAParametersPlus(std::shared_ptr<const Telegram> telegram);
+    void process_UBAParameterWWPlus(std::shared_ptr<const Telegram> telegram);
     void process_UBAOutdoorTemp(std::shared_ptr<const Telegram> telegram);
     void process_UBASetPoints(std::shared_ptr<const Telegram> telegram);
     void process_UBAFlags(std::shared_ptr<const Telegram> telegram);
@@ -168,6 +174,8 @@ class Boiler : public EMSdevice {
     void set_warmwater_circulation(const char * value, const int8_t id);
     void set_warmwater_temp(const char * value, const int8_t id);
     void set_flow_temp(const char * value, const int8_t id);
+    void set_heating_activated(const char * value, const int8_t id);
+    void set_heating_temp(const char * value, const int8_t id);
     void set_min_power(const char * value, const int8_t id);
     void set_max_power(const char * value, const int8_t id);
     void set_hyst_on(const char * value, const int8_t id);